[Fix] 킥보드 페이지에서 다른 사용자가 등록한 킥보드는 모달이 안올라오게 하기

+++ TeamProject/TeamProject/View/KickBoardRegisterView/KickBoardRegisterViewController.swift
@@ -191,6 +191,12 @@ extension KickBoardRegisterViewController {
             print("Record not found for tapped POI")
             return
         }
+        
+        let id = UserManager.shared.getUser()?.id
+        if record.userID != id {
+            return
+        }
+        
         print("Tapped record: \(record)")
         
         let vc = DeleteModalViewController(
